+++ README.md
@@ -28,7 +28,7 @@ Latest API Docs is <a target="_blank" href="https://wso2-extensions.github.io/si
 
 ```
      <dependency>
-        <groupId>org.wso2.extension.siddhi.io.tcp</groupId>
+        <groupId>io.siddhi.extension.io.tcp</groupId>
         <artifactId>siddhi-io-tcp</artifactId>
         <version>x.x.x</version>
      </dependency>
