--- component/src/main/java/org/wso2/extension/siddhi/io/tcp/transport/utils/BinaryMessageConverterUtil.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied. See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-package org.wso2.extension.siddhi.io.tcp.transport.utils;
-
-import io.netty.buffer.ByteBuf;
-
-import java.io.UnsupportedEncodingException;
-
-/**
- * Util helping to convert from Siddhi Event to byte message.
- */
-public final class BinaryMessageConverterUtil {
-
-    private BinaryMessageConverterUtil() {
-
-    }
-
-    public static String getString(ByteBuf byteBuffer, int size) throws UnsupportedEncodingException {
-        byte[] bytes = new byte[size];
-        byteBuffer.readBytes(bytes);
-        return new String(bytes, Constant.DEFAULT_CHARSET);
-    }
-}
