--- component/src/main/java/org/wso2/extension/siddhi/io/tcp/transport/utils/FlowController.java
@@ -1,64 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied. See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-
-package org.wso2.extension.siddhi.io.tcp.transport.utils;
-
-import java.util.concurrent.locks.Condition;
-import java.util.concurrent.locks.ReentrantLock;
-
-/**
- * Flow Controller to control consuming.
- */
-public class FlowController {
-
-    private volatile boolean paused = false;
-    private ReentrantLock lock;
-    private Condition condition;
-
-    public FlowController() {
-        this.lock = new ReentrantLock();
-        this.condition = lock.newCondition();
-    }
-
-    public void pause() {
-        paused = true;
-    }
-
-    public void resume() {
-        paused = false;
-        try {
-            lock.lock();
-            condition.signalAll();
-        } finally {
-            lock.unlock();
-        }
-    }
-
-    public void barrier() {
-        if (paused) { //spurious wakeup condition is deliberately traded off for performance
-            lock.lock();
-            try {
-                condition.await();
-            } catch (InterruptedException ie) {
-                Thread.currentThread().interrupt();
-            } finally {
-                lock.unlock();
-            }
-        }
-    }
-}
