+++ component/src/test/resources/log4j.properties
@@ -31,6 +31,6 @@ log4j.appender.stdout.layout.ConversionPattern=%m%n
 log4j.logger.com.zaxxer.hikari=WARN
 log4j.logger.com.zaxxer.hikari.pool=WARN
 
-log4j.appender.testNG=org.wso2.extension.siddhi.io.tcp.utils.LoggerAppender
+log4j.appender.testNG=io.siddhi.extension.io.tcp.utils.LoggerAppender
 log4j.appender.testNG.layout=org.apache.log4j.EnhancedPatternLayout
 log4j.appender.testNG.layout.ConversionPattern=%d{ABSOLUTE} %5p %c{1}:%L - %m%
