+++ component/src/test/resources/testng.xml
@@ -22,8 +22,8 @@
 <suite name="Tcp-IO-Test-Suite">
     <test name="tcp-tests" enabled="true" preserve-order="true" parallel="false">
         <classes>
-            <class name="org.wso2.extension.siddhi.io.tcp.TCPSinkTestCase"/>
-            <class name="org.wso2.extension.siddhi.io.tcp.TCPSourceTestCase"/>
+            <class name="io.siddhi.extension.io.tcp.TCPSinkTestCase"/>
+            <class name="io.siddhi.extension.io.tcp.TCPSourceTestCase"/>
         </classes>
     </test>
 </suite>
